extractor_service/requirements.txt
fastapi
uvicorn
langextract
python-multipart
pdfplumber
python-docx